clean esdrive
1. clean esdrive
2. load svm model in windows system
