+++ core/src/main/assembly/zipSVMWithSGDModel.xml
@@ -19,7 +19,7 @@
     <id>zipSVMWithSGDModel</id>
     <baseDirectory>${svmSgdModel.value}</baseDirectory>
     <formats>
-        <format>tar.gz</format>
+        <format>zip</format>
     </formats>
     <fileSets>
         <fileSet>
