+++ core/src/main/java/gov/nasa/jpl/mudrod/main/MudrodEngine.java
@@ -36,6 +36,8 @@ import java.nio.file.Files;
 import java.nio.file.attribute.PosixFilePermissions;
 import java.util.List;
 import java.util.Properties;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
 
 /**
  * Main entry point for Running the Mudrod system. Invocation of this class is
@@ -174,43 +176,56 @@ public class MudrodEngine {
 
   private String decompressSVMWithSGDModel(String archiveName) throws IOException {
 
-    URL scmArchive = getClass().getClassLoader().getResource(archiveName);
-    if (scmArchive == null) {
-      throw new IOException("Unable to locate " + archiveName + " as a classpath resource.");
-    }
-    File tempDir = Files.createTempDirectory("mudrod", PosixFilePermissions.asFileAttribute(PosixFilePermissions.fromString("rwxrw----"))).toFile();
-    File archiveFile = new File(tempDir, archiveName);
-    FileUtils.copyURLToFile(scmArchive, archiveFile);
+		URL scmArchive = getClass().getClassLoader().getResource(archiveName);
+		System.out.println("scmArchive"  +scmArchive.toString());
+		if (scmArchive == null) {
+			throw new IOException("Unable to locate " + archiveName + " as a classpath resource.");
+		}
+		// File tempDir = Files.createTempDirectory("mudrod",
+		// PosixFilePermissions.asFileAttribute(PosixFilePermissions.fromString("rwxrw----"))).toFile();
+		File tempDir = Files.createTempDirectory("mudrod").toFile();
+		assert tempDir.setWritable(true);
+		File archiveFile = new File(tempDir, archiveName);
+		FileUtils.copyURLToFile(scmArchive, archiveFile);
 
-    // Decompress archive
-    int BUFFER_SIZE = 512000;
-    GzipCompressorInputStream gzipIn = new GzipCompressorInputStream(new FileInputStream(archiveFile));
-    try (TarArchiveInputStream tarIn = new TarArchiveInputStream(gzipIn)) {
-      TarArchiveEntry entry;
+		// Decompress archive
+		int BUFFER_SIZE = 512000;
+		System.out.println(archiveName);
+		ZipInputStream zipIn = new ZipInputStream(new FileInputStream(archiveFile));
+		ZipEntry entry;
+		while ((entry = zipIn.getNextEntry()) != null) {
+			// If the entry is a directory, create the directory.
+			if (entry.isDirectory()) {
+				File f = new File(tempDir, entry.getName());
+				boolean created = f.mkdirs();
+				if (!created) {
+					LOG.error("Unable to create directory '%s', during extraction of archive contents.",
+							f.getAbsolutePath());
+				}
+			} 
+			
+		}
+		
+		zipIn = new ZipInputStream(new FileInputStream(archiveFile));
+		while ((entry = zipIn.getNextEntry()) != null) {
+			// If the entry is a directory, create the directory.
+			if (!entry.isDirectory()) {
+				System.out.println("3");
+				System.out.println(tempDir+ "/"+ entry.getName());
+				int count;
+				byte data[] = new byte[BUFFER_SIZE];
+				FileOutputStream fos = new FileOutputStream(new File(tempDir, entry.getName()), false);
+				try (BufferedOutputStream dest = new BufferedOutputStream(fos, BUFFER_SIZE)) {
+					while ((count = zipIn.read(data, 0, BUFFER_SIZE)) != -1) {
+						dest.write(data, 0, count);
+					}
+				}
+			}
+		}
 
-      while ((entry = (TarArchiveEntry) tarIn.getNextEntry()) != null) {
-        // If the entry is a directory, create the directory.
-        if (entry.isDirectory()) {
-          File f = new File(tempDir, entry.getName());
-          boolean created = f.mkdir();
-          if (!created) {
-            LOG.error("Unable to create directory '%s', during extraction of archive contents.", f.getAbsolutePath());
-          }
-        } else {
-          int count;
-          byte data[] = new byte[BUFFER_SIZE];
-          FileOutputStream fos = new FileOutputStream(new File(tempDir, entry.getName()), false);
-          try (BufferedOutputStream dest = new BufferedOutputStream(fos, BUFFER_SIZE)) {
-            while ((count = tarIn.read(data, 0, BUFFER_SIZE)) != -1) {
-              dest.write(data, 0, count);
-            }
-          }
-        }
-      }
-    }
+		return new File(tempDir, StringUtils.removeEnd(archiveName, ".zip")).toURI().toString();
+	}
 
-    return new File(tempDir, StringUtils.removeEnd(archiveName, ".tar.gz")).toURI().toString();
-  }
 
   /**
    * Preprocess and process various
